pyserver/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
torch
torchvision
pillow
